haloAdd project files.

+++ Lib/Entity/InvoiceDetail.cs
@@ -15,8 +15,8 @@ namespace Lib.Entity
         [Key]
         public int ID_Food { get; set; }
         public int Quantity { get; set; }
-        public float Price { get; set; }
-        public float Total_Money { get; set; }
+        public double Price { get; set; }
+        public double Total_Money { get; set; }
 
     }
 }
